Auto increment build numbers
The version file generator will now read the previous version file and determine what
the previous build's MAJOR.MINOR.MICRO.BUILD version is. It will then compare this to
the current MAJOR.MINOR.MICRO version of the source code.
If the two are the same, the build number will be incremented.
Otherwise, it gets reset to 0