Only save unique snapshots in history.
Every update, the snapshot will squash identical snapshots into each other.
If the most recent snapshot is the same as the previous, it is discarded.
If the most recent snapshot is the same but lacks some of the windows the
previous has (ie: they have been closed), it is discarded.
If the most recent snapshot has all of the same windows as the previous
with some extras (ie: new windows have been opened) then the previous (lesser)
snapshot is discarded.