Add basic support for rules.
In your `history.json`, you can now define rules that will apply to windows that aren't in your snapshot.
The rules will match either a window's name or executable. It will try an exact match and if that doesn't work,
it will compile the rule as a regex and see if that matches.

Currently there is no user friendly way to define rules, so I'll have to work on that in the future