Parse snapshot file to dataclasses, use type hints.
It was getting pretty confusing remembering what my variables were, how this particular dict was laid out,
whether this dict key always exists or might not. So I created some dataclasses that the snapshot file
will be automatically parsed into and some types to go along with it.

This makes the intellisense in my IDE much more useful