Enable main loop dynamic sleep.
Previously, if you changed the snapshot frequency while the main loop was sleeping, it would continue
sleeping for however long the previous sleep value was. So if it had just started an hour sleep and
you changed the frequency to 1 minute, it would not exit the sleep until the hour was up.

Now, it will measure the time elapsed since it started sleeping. If it is greater than the current
sleep value, the sleep will be exited