Add ability to manage "global" rules in GUI.
I'm not yet sure what "global" means, as I haven't implemented the backend for this.
Ideally, I will be able to figure out whether a rule is compatible with a certain display configuration and apply it.
Perhaps I can simply calculate whether the rule's RECT fits entirely within the RECT of the current display configuration, and apply it if so.

In any case, this will be a pretty cool feature once I have that figured out.
Hopefully, in the future, I'll be able to add display matching to specific rules, so the user could say
"match this rule with every display that has a name 'Dell Ultrawide.*'".

That will require some dataclass reshuffling