Add test for the installer.
The test will run the installer a few times, validating that it does install files, overwrite existing installs
and removes files when uninstalled. It also checks that the new `/StartAfterInstall` flag works.

I don't recommend running the test installer locally, just because it will mess with your existing install.
It will run automatically as part of the release workflow