Don't resize "unresizeable" windows.
Certain windows are not meant to be resized and sometimes it's for good reasons.
The "sound" dialog in Windows (right click taskbar sound icon -> "sounds") is OLD and simply does not
allow resizing because all of it's widgets are absolutely positioned.
Similarly, the reminders window that Outlook spawns does not resize well.

Now, when applying snapshots or rules, these windows will be detected. They will be moved to the coords
that the user requests but they will not be resized