Fix TypeError when matching rules against a window.
The error would occur when two unequal rules matched a window with the same number of points.
EG: two identical rules but one with a different `rule_name`. `find_matching_rules` failed to
supply a sort key, meaning that when the points were the same the dataclasses would then be
compared, which would lead to a `<` comparison between them, raising a TypeError.

The error was pretty rare, only happening to me once since the code was comitted.