Fix windows on inactive virtual desktops being marked as invalid.
Validity check was whether the window was "cloaked" by the shell, which is the case for valid windows
on other virtual desktop. New filter uses `pyvda`, a super cool python library for working with windows
virtual desktops.

Now, we check we can fetch a valid desktop GUID for the relevant window. If we can, then the window is good.

Currently there is no support for making rules that match windows per virtual desktop, or that moves windows
to other virtual desktops, although this will likely follow soon