Fix `COMError` importing `pyvda` in CI tests.
Issue appears to be that the build of windows used by github actions does not have the `IVirtualDesktopManagerInternal2` interface
that pyvda requires for Windows builds under 21313. Not sure if this is a bug or a discrepancy between the server and consumer Windows builds.

The patch is an absolutely filthy hack. The test suite will detect if it's running in Github Actions and will re-write the `pyvda.utils`
submodule source file so that the `IVirtualDesktopManagerInternal2` object is never requested.
Truly vile indeed.