Add systray option to "rescue" windows.
I realised that if the user directly applies a layout or a rule, there are no checks for if that rule is valid or not.
This means that you can accidentally send windows off-screen, making it really hard to get them back.

Previously, you could use `ALT+SPACE M` and use arrow keys to move it, create another rule to move the window back or
restore your current snapshot. Now, you can just open the systray menu, click "Rescue windows" and it will find any off-screen windows
and move them back to (0, 0) coords