Add cached factory for JSONFile instances.
The use case is that the program settings are program wide and many different modules will need access to those settings.
Creating a singleton class is more work than it should be so I've added a cached function that just initialises `JSONFile`s
for you.

Now, any part of the program can do `load_json('settings')` to get the settings file instance without me needing to pass it around.