Add settings to control how window history is kept.
Added settings to control whether dead windows are deleted, whether to delete snapshots older than X timespan and the maximum number of snapshots to keep.
The reason for adding this is because of the new on window spawn features. It wasn't working very well previously because it would only remember the LKP of
processes that are still alive. Processes with no windows would be forgotten relatively quickly.

Now, the user can disable forgetting dead windows but also how long dead windows should be kept around, which will prevent the `history.json` file from
growing out of control.