Fix bug where unresizeable windows would be resized.
Main example of this was the Outlook reminders prompt that kept being ballooned to a size and shape that didn't make sense.
Turns out I was adding the current w/h to the bottom right coord of the new rect, rather than top left. This meant it would be set to
LKP size PLUS current size.