Fix some windows being moved after LKP applied.
Problem was with placement. If we set the placement to that of the last window (eg: maximised), when
we called `win32gui.ShowWindow` we would pass `SW_SHOWNORMAL`, which would, from MS docs:
> If the window is minimized, maximized, or arranged, the system restores it to its original size and position

There's no way to call show window and just activate/raise the window as far as I can see so the solution is
to check the placement of the window to see if it's minimised and then manually set a restored placement.