Add option to skip non-resizable windows when running on-spawn service.
Some programs (eg: winword) put up a splash screen before launch. This is moved and then added to the current snapshot before
the full window spawns. Then, when the actual window appears it is resized to match the splash screen, since that's now the LKP.

This is incredibly frustrating, so I added an option to completely skip non-resizable windows in the on-spawn service. This means
they won't be moved, resized or added to the snapshot.

I have yet to test this with things like MS teams notifications