Add option to filter LKP by resizability.
As mentioned in a previous commit, some programs have splash screens. That commit dealt with it by ignoring non-resizable windows.
This commit adds the ability to treat them differently to resizable windows.

When enabled, LKPs will be filtered based on whether their resizability is the same as the current window.