Add workaround for pyinstaller bootloader not propagating signals to child.
Usually, pyinstaller tries to pass signals (eg: SIGINT, SIGTERM) to the child so that it can properly respond.
When running in `--onefile` mode (ie: bundled as single exe file), this doesn't happen. It appears that the bootloader
spawns the child as a separate process that runs completely separately. Kill signals sent to the parent do not get
forwarded to the child, which keeps running long after the parent ends.

To work around this, I've added a check to see if a parent process exists with the same name and exe. The app will
regularly check if the parent is still running. If not, the app exits.

I also did some tidying up of the shutdown logic. I kept running into "memory at 0xFFFFF... cannot be referenced"
errors popping up on Windows shutdown. I suspect these were due to exiting the wx mainloop without cleaning up
resources (ie: destroying the app). I've made it so that `shutdown` is called after the event loop exits and it
destroys the wxapp as part of the process